fix(babel): Pass chunk filename to output plugin transform
This appears to be an oversight, as the filename is passed at the same
place in the input plugin. Without this, a Babel error including text
"Preset /* your preset */ requires a filename to be set when babel is
called directly" is thrown when using a custom preset.

The test is not ideal, but without replicating a huge stack of other
Rollup plugins and complicated Babel config, it's unclear how else to
regress this. A local edit of the Rollup Babel plugin in node_modules
of the offending build does indicate the error is mitigated, however.