feat(swc): sets filename property, so to make additional swc features… (#1761)
feat(swc): sets filename property, so to make additional swc features available

Background: When using the swc rollup plugin in vitest, debugging will not work. I assume this comes from the filename property not being set, as the documentation states:
```
The filename is optional, but not all of Swc's
functionality is available when the filename is unknown, because a subset of options rely on the filename for their functionality.
```

Setting a filename should not have any side effects, just make debugging and other features work.