feat(pluginutils): add `createHookFilter()`
When trying to port a plugin that uses `createFilter()` to hook filters
it's hard keep the functionality 100% the same, as `createFilter()` does
various pre-processing, like normalising paths and, resolving relative
paths, exluding paths which contain \0, etc.

To make porting easier, provide a `createHookFilter()` which instead of
returning a function, returns a filter object that can be assigned
to "filter.id" or "filter.code", and behaves the same as if createFilter()
was used and the returned function used in a hook.

To make sure that the implementation of `createHookFilter()` matches
`createFilter()`, `createFilter()` now calls `createHookFilter()`
internally and uses the same object to create the test function.