feat(eslint): add support for ESLint v9 (#1905)
* chore(repo): add explicit eslint dependency

The root workspace was missing an eslint dependency, but was
invoking the eslint CLI directly in the scripts. Make the
dependency explicit.

* fix(eslint): switch flat config to .mjs in test

The parent package hasn't opt in to ES modules, which breaks on
Node.js v18:

    (node:5286) Warning: To load an ES module, set "type": "module" in the package.json or use the .mjs extension.

* feat(eslint): add support for ESLint v9

Update package.json to make eslint a peerDependency, so that the
user has control over which version to pick. Move the existing
eslint dependency to devDependencies so that tests work as expected.

Introduce a new test to ensure the plugin works with ESLint v9. To
do so without breaking v8 tests, install ESLint v9 under a different
alias, and use an ESM import mocking library to override the "eslint"
import.

The current directory needs to be changed during the ESLint v9 test
so that ESLint finds its config file properly. This requires ava
workerThreads to be disabled:

    TypeError [ERR_WORKER_UNSUPPORTED_OPERATION]: process.chdir() is not supported in workers